implement named primitives consistently; restructure module layout
