+++ src/callable/primitive/mod.rs
@@ -1,0 +1,14 @@
+mod c;
+pub use c::*;
+
+mod callstack;
+pub use callstack::*;
+
+mod list;
+pub use list::*;
+
+mod paste;
+pub use paste::*;
+
+mod q;
+pub use q::*;
