+++ src/lib.rs
@@ -6,6 +6,6 @@ pub mod error;
 pub mod lang;
 pub mod parser;
 
-pub mod r_builtins;
-pub mod r_repl;
-pub mod r_vector;
+pub mod callable;
+pub mod repl;
+pub mod vector;
