--- src/r_builtins/mod.rs
@@ -1,7 +1,0 @@
-/// builtins
-///
-/// Everything builtins related lives here!
-///
-pub mod builtins;
-pub mod paste;
-pub mod callstack;
