--- src/r_repl/mod.rs
@@ -1,8 +1,0 @@
-/// R REPL
-///
-/// Everything REPL related lives here!
-///
-pub mod highlight;
-pub mod prompt;
-pub mod repl;
-pub mod validator;
