+++ src/repl/highlight.rs
