+++ src/repl/mod.rs
@@ -1,0 +1,7 @@
+pub mod highlight;
+pub mod prompt;
+
+mod repl;
+pub use repl::*;
+
+pub mod validator;
