+++ src/vector/coercion.rs
